cleaned up code so all files match same coding patterns.
