+++ ui/app/(tabs)/_layout.tsx
@@ -1,11 +1,11 @@
 import { Tabs } from 'expo-router';
 import React from 'react';
 
-import { Icon } from '@/common/Icon';
+import  Icon from '@/common/Icon';
 import { Colors } from '@/constants/Colors';
 import { useColorScheme } from '@/hooks/useColorScheme';
 
-export default function TabLayout() {
+const TabLayout = () => {
   const colorScheme = useColorScheme();
 
   return (
@@ -35,3 +35,5 @@ export default function TabLayout() {
     </Tabs>
   );
 }
+
+export default TabLayout;
