+++ ui/app/+html.tsx
@@ -5,7 +5,9 @@ import { type PropsWithChildren } from 'react';
  * This file is web-only and used to configure the root HTML for every web page during static rendering.
  * The contents of this function only run in Node.js environments and do not have access to the DOM or browser APIs.
  */
-export default function Root({ children }: PropsWithChildren) {
+const  Root = ({ 
+  children 
+}: PropsWithChildren) => {
   return (
     <html lang="en">
       <head>
@@ -37,3 +39,5 @@ body {
     background-color: #000;
   }
 }`;
+
+export default Root;
