+++ ui/app/+not-found.tsx
@@ -1,19 +1,19 @@
 import { Link, Stack } from 'expo-router';
 import { StyleSheet } from 'react-native';
 
-import { ThemedText } from '@/common/Typography';
-import { ThemedView } from '@/common/Card';
+import Typography from '@/common/Typography';
+import Card from '@/common/Card';
 
-export default function NotFoundScreen() {
+const NotFoundScreen = () => {
   return (
     <>
       <Stack.Screen options={{ title: 'Oops!' }} />
-      <ThemedView style={styles.container}>
-        <ThemedText type="title">This screen doesn't exist.</ThemedText>
+      <Card style={styles.container}>
+        <Typography type="title">This screen doesn't exist.</Typography>
         <Link href="/" style={styles.link}>
-          <ThemedText type="link">Go to home screen!</ThemedText>
+          <Typography type="link">Go to home screen!</Typography>
         </Link>
-      </ThemedView>
+      </Card>
     </>
   );
 }
@@ -30,3 +30,5 @@ const styles = StyleSheet.create({
     paddingVertical: 15,
   },
 });
+
+export default NotFoundScreen;
