+++ ui/common/Card/index.tsx
@@ -1,19 +1,3 @@
-import { View, type ViewProps } from 'react-native';
+import Card from "./Card";
 
-import { useThemeColor } from '@/hooks/useThemeColor';
-
-export type ICarProps = ViewProps & {
-  lightColor?: string;
-  darkColor?: string;
-};
-
-export function Card({ 
-  style, 
-  lightColor, 
-  darkColor, 
-  ...otherProps }: ICarProps
-) {
-  const backgroundColor = useThemeColor({ light: lightColor, dark: darkColor }, 'background');
-
-  return <View style={[{ backgroundColor }, style]} {...otherProps} />;
-}
+export default Card;
