+++ ui/common/InputText/index.tsx
@@ -1,24 +1,3 @@
-import { TextInput, type TextInputProps, StyleSheet } from 'react-native';
+import InputText from "./InputText";
 
-export type IInputTextProps = TextInputProps & {
-  placeholder?: string
-}
-
-export function InputText({
-  placeholder,
-  style,
-  ...rest
-}: IInputTextProps) {
-  return (
-    <TextInput placeholder={placeholder} style={styles.input} {...rest} />
-  );
-};
-
-const styles = StyleSheet.create({
-  input: {
-    height: 40,
-    margin: 12,
-    borderWidth: 1,
-    padding: 10,
-  },
-});
+export default InputText;
