+++ ui/common/Radio/index.tsx
@@ -1,17 +1,3 @@
-import BouncyCheckbox from "react-native-bouncy-checkbox";
-import Typography from "../Typography";
+import RadioButton from "./Radio";
 
-export type IRadioProps = {
-  text: string,
-}
-
-export const RadioButton = ({
-  text,
-}:IRadioProps) => {
-  return (
-    <Typography>
-      <BouncyCheckbox text={text} onPress={(isChecked: boolean) => {}} />
-      {text}
-    </Typography>
-  )
-}
+export default RadioButton;
